Changes to existing viewer code.
Add hostDomain and isCrossDomain attributes onto viewer object.

The isCrossDomain flag is consulted later by the construction helper to decide whether to render the editor inside the sidebar.  The hostDomain is used as the target by the login functionality.

There is also quite a few places in the codebase that were calling:  "parseInt(this.api.getId(), 10)".  The api method getModelId does the exact same thing.  When I've found them I'm replacing them to stay DRY