Api support for logout
This adds a login & logout methods to the editor, and has the helper
that listens to click events call them.  The login class handles
co-ordination with remote socket to kick off the call once it's shown.

The RemoteSocket class is gets a new RPC endpoint, 'logout' and is
refactored to always dispatch whatever data it receives onto the
models.  This simplifies the other end, as it can just call
'loggedInStatus' with the appropriate data and the backbone models will
update accordingly.